Finished refactoring of FeatureChunkCache.
The change was about making 1 adapter to have 1 FeatureChunkCache, and
this, to have also just 1 IndexedDBStore.

As one track only has 1 adapter, this implies the adapter to manage
several categories in the case it manages several subtracks.

In the current implementation, the categories are ObjectStores in
IndexedDB.

By the moment, the refactor has not been done in the other tracks.